+++ examples/4. rate-limit/index.ts
@@ -10,8 +10,8 @@ new Spider({
     handlers: [{
         pattern: '**',
         handle: async function (response) {
-            let title = await response.xpath('//h1') as string;
-            logger.info(title as string);
+            let title = await response.xpath('//h1');
+            logger.info(title.text());
             this.enqueue('https://httpbin.org/html');
         },
     }],